Rename IModifedCheckoutObject to CheckoutObjectWithUser, test cases update to match new return format
